update exercise & test

+++ exercise-errors.go
@@ -10,7 +10,7 @@ func Sqrt(x float64) (float64, error) {
 	z := float64(1)
 	s := float64(0)
 	if x < 0 {
-		return float64(math.NaN()),ErrNegativeSqrt(x)
+		return math.NaN(),ErrNegativeSqrt(x)
 	}else if x == 0 {
 		return 0,nil
 	}else {
@@ -23,7 +23,7 @@ func Sqrt(x float64) (float64, error) {
 }
 
 func (e ErrNegativeSqrt) Error() string {
-	return fmt.Sprintf("cannot Sqrt negative number: %v",e)
+	return fmt.Sprintf("cannot Sqrt negative number: %v",float64(e))
 }
 func main() {
 	fmt.Println(Sqrt(-2))
